Move code files to src
